function/requirements.txt
looker-sdk==24.20.0
google-cloud-aiplatform==1.56.0
google-cloud-logging==3.11.3
pydantic==2.7.4
google-cloud-storage==2.19.0
Flask==3.1.0
Jinja2==3.1.5